refactor: dedup formatDuration/formatTime into shared formatters util
Byte-identical ns-to-m:ss formatting logic was reimplemented locally
in 7 components (5 as formatDuration, 2 as formatTime). Consolidated
into formatDuration in src/lib/utils/formatters.ts and updated all
call sites; the two formatTime call sites were renamed to
formatDuration for naming consistency across the codebase.

Part of the #577 code-quality roadmap.
