+++ src/lib/components/AlbumDetailView.svelte
@@ -6,6 +6,7 @@
   import { playerStore } from "../stores/player.svelte";
   import { playlistsStore } from "../stores/playlists.svelte";
   import { shuffleArray } from "../utils/shuffle";
+  import { formatDuration } from "../utils/formatters";
   import CoverArt from "./CoverArt.svelte";
   import SongRating from "./SongRating.svelte";
   import FavouriteCornerFlag from "./FavouriteCornerFlag.svelte";
@@ -421,14 +422,6 @@
       })
       .catch((err) => console.error(err))
       .finally(() => loading = false);
-  }
-
-  function formatDuration(ns: number | undefined): string {
-    if (!ns) return "0:00";
-    const sec = Math.floor(ns / 1_000_000_000);
-    const m = Math.floor(sec / 60);
-    const s = sec % 60;
-    return `${m}:${s < 10 ? "0" : ""}${s}`;
   }
 
   async function rateSong(song: Song, rating: number) {
