+++ src/lib/components/ArtistDetailView.svelte
@@ -4,6 +4,7 @@
   import { playerStore } from "../stores/player.svelte";
   import { playlistsStore } from "../stores/playlists.svelte";
   import { shuffleArray } from "../utils/shuffle";
+  import { formatDuration } from "../utils/formatters";
   import CoverArt from "./CoverArt.svelte";
   import CoverStack from "./CoverStack.svelte";
   import AlbumCard from "./AlbumCard.svelte";
@@ -96,14 +97,6 @@
     collectionStore.refreshLibrary();
     tagsStore.load();
     refetchSongs();
-  }
-
-  function formatDuration(ns: number | undefined): string {
-    if (!ns) return "0:00";
-    const sec = Math.floor(ns / 1_000_000_000);
-    const m = Math.floor(sec / 60);
-    const s = sec % 60;
-    return `${m}:${s < 10 ? "0" : ""}${s}`;
   }
 
   async function rateSingle(song: Song, rating: number) {
